# swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/validator/:stash_account": {
      "get": {
        "description": "Returns most recent aggreate information for validator account, as well as lists of validator era sequences for last eras_limit eras,\nvalidator sessio sequences for last sessions_limit sessions, and most recent delegations.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Gets aggregate details for validator",
        "operationId": "getValidatorByStash",
        "parameters": [
          {
            "type": "string",
            "description": "StashAccount",
            "name": "StashAccount",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "description": "SessionsLimit",
            "name": "SessionsLimit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "description": "ErasLimit",
            "name": "ErasLimit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ValidatorAggregateView"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          }
        }
      }
    },
...
    "/validators/for_min_height/:height": {
      "get": {
        "description": "Returns list of most recent aggreate information for all validator accounts that have been validators after provided min height",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Gets all validator aggregate information after height",
        "operationId": "getValidatorForMinHeight",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "Height",
            "name": "Height",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ValidatorAggregatesView"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          }
        }
      }
    },
    "/validators_summary": {
      "get": {
        "description": "Returns validator summaries for all accounts (or single account if stash_account query is provided)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Gets all validators for height",
        "operationId": "getValidatorSummary",
        "parameters": [
          {
            "type": "string",
            "description": "Interval",
            "name": "Interval",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "Period",
            "name": "Period",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "StashAccount",
            "name": "StashAccount",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ValidatorsSummariesView"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Delegation": {
      "type": "object",
      "properties": {
        "amount": {
          "$ref": "#/definitions/Quantity"
        },
        "era": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Era"
        },
        "validator_stash": {
          "type": "string",
          "x-go-name": "ValidatorStash"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/usecase/common"
    },
...
    "ID": {
      "type": "integer",
      "format": "int64",
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/types"
    },
...
    "Quantity": {
      "description": "Quantity is a wrapper for big.Int",
      "type": "object",
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/types"
    },
...
    "StringArray": {
      "type": "array",
      "title": "StringArray represents a one-dimensional array of the PostgreSQL character types.",
      "items": {
        "type": "string"
      },
      "x-go-package": "github.com/lib/pq"
    },
...
    "Time": {
      "type": "object",
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/types"
    },
...
    "ValidatorAgg": {
      "type": "object",
      "properties": {
        "accumulated_uptime": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AccumulatedUptime"
        },
        "accumulated_uptime_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AccumulatedUptimeCount"
        },
        "created_at": {
          "$ref": "#/definitions/Time"
        },
        "display_name": {
          "type": "string",
          "x-go-name": "DisplayName"
        },
        "id": {
          "$ref": "#/definitions/ID"
        },
        "recent_as_validator_height": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RecentAsValidatorHeight"
        },
        "recent_at": {
          "$ref": "#/definitions/Time"
        },
        "recent_at_height": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RecentAtHeight"
        },
        "started_at": {
          "$ref": "#/definitions/Time"
        },
        "started_at_height": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StartedAtHeight"
        },
        "stash_account": {
          "type": "string",
          "x-go-name": "StashAccount"
        },
        "updated_at": {
          "$ref": "#/definitions/Time"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/model"
    },
    "ValidatorEraSeq": {
      "type": "object",
      "properties": {
        "commission": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Commission"
        },
        "controller_account": {
          "type": "string",
          "x-go-name": "ControllerAccount"
        },
        "end_height": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "EndHeight"
        },
        "era": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Era"
        },
        "id": {
          "$ref": "#/definitions/ID"
        },
        "index": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Index"
        },
        "own_stake": {
          "$ref": "#/definitions/Quantity"
        },
        "reward_points": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RewardPoints"
        },
        "session_accounts": {
          "$ref": "#/definitions/StringArray"
        },
        "stakers_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StakersCount"
        },
        "stakers_stake": {
          "$ref": "#/definitions/Quantity"
        },
        "start_height": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StartHeight"
        },
        "stash_account": {
          "type": "string",
          "x-go-name": "StashAccount"
        },
        "time": {
          "$ref": "#/definitions/Time"
        },
        "total_stake": {
          "$ref": "#/definitions/Quantity"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/model"
    },
    "ValidatorSessionSeq": {
      "type": "object",
      "properties": {
        "end_height": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "EndHeight"
        },
        "id": {
          "$ref": "#/definitions/ID"
        },
        "online": {
          "type": "boolean",
          "x-go-name": "Online"
        },
        "session": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Session"
        },
        "start_height": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StartHeight"
        },
        "stash_account": {
          "type": "string",
          "x-go-name": "StashAccount"
        },
        "time": {
          "$ref": "#/definitions/Time"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/model"
    },
    "ValidatorSummaryRow": {
      "type": "object",
      "properties": {
        "commission_avg": {
          "type": "number",
          "format": "double",
          "x-go-name": "CommissionAvg"
        },
        "commission_max": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CommissionMax"
        },
        "commission_min": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CommissionMin"
        },
        "own_stake_avg": {
          "$ref": "#/definitions/Quantity"
        },
        "own_stake_max": {
          "$ref": "#/definitions/Quantity"
        },
        "own_stake_min": {
          "$ref": "#/definitions/Quantity"
        },
        "reward_points_avg": {
          "type": "number",
          "format": "double",
          "x-go-name": "RewardPointsAvg"
        },
        "reward_points_max": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RewardPointsMax"
        },
        "reward_points_min": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RewardPointsMin"
        },
        "stakers_count_avg": {
          "type": "number",
          "format": "double",
          "x-go-name": "StakersCountAvg"
        },
        "stakers_count_max": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StakersCountMax"
        },
        "stakers_count_min": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StakersCountMin"
        },
        "stakers_stake_avg": {
          "$ref": "#/definitions/Quantity"
        },
        "stakers_stake_max": {
          "$ref": "#/definitions/Quantity"
        },
        "stakers_stake_min": {
          "$ref": "#/definitions/Quantity"
        },
        "time_bucket": {
          "type": "string",
          "x-go-name": "TimeBucket"
        },
        "time_interval": {
          "type": "string",
          "x-go-name": "TimeInterval"
        },
        "total_stake_avg": {
          "$ref": "#/definitions/Quantity"
        },
        "total_stake_max": {
          "$ref": "#/definitions/Quantity"
        },
        "total_stake_min": {
          "$ref": "#/definitions/Quantity"
        },
        "uptime_avg": {
          "type": "number",
          "format": "double",
          "x-go-name": "UptimeAvg"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/store"
    },
...
    "syncableInfo": {
      "type": "object",
      "properties": {
        "time": {
          "$ref": "#/definitions/Time"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/usecase/validator"
    }
  },
  "responses": {
...
    "BadRequestResponse": {
      "description": "",
      "schema": {
        "type": "object"
      },
      "headers": {
        "error": {},
        "status": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "ValidatorAggregateView": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Time"
      },
      "headers": {
        "created_at": {},
        "delegations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Delegation"
          }
        },
        "display_name": {
          "type": "string"
        },
        "id": {},
        "last_era_sequences": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidatorEraSeq"
          }
        },
        "last_session_sequences": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidatorSessionSeq"
          }
        },
        "recent_as_validator_height": {
          "type": "integer",
          "format": "int64"
        },
        "recent_at": {},
        "recent_at_height": {
          "type": "integer",
          "format": "int64"
        },
        "started_at": {},
        "started_at_height": {
          "type": "integer",
          "format": "int64"
        },
        "stash_account": {
          "type": "string"
        },
        "updated_at": {},
        "uptime": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "ValidatorAggregatesView": {
      "description": "",
      "headers": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidatorAgg"
          }
        }
      }
    },
    "ValidatorsSummariesView": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/syncableInfo"
      },
      "headers": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidatorSummaryRow"
          }
        },
        "last_indexed_era": {},
        "last_indexed_session": {}
      }
    },
...
  }
}